Returns error when vpc-ipv4-cidr-blocks is empty
When a new interface is added there is a lag where the ec2 metadata
service returns incomplete information. The logic that accomodates
this depends on the aws.GetInterfaces() method to return an error
to force it to retry. This change checks the length of list of
vpc-ipv4-cidr-blocks and generates an error if it is empty. This
will allow the interface setup logic to retry and avoid pods with
incorrect routes